lastminute/pyproject.toml
[project]
name = "lastminute"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.24.0",
    "aiomysql>=0.3.2",
    "apscheduler>=3.11.2",
    "beautifulsoup4>=4.14.3",
    "cryptography>=46.0.3",
    "greenlet>=3.3.0",
    "levenshtein>=0.27.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.45",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
